requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.30.0
cryptography==38.0.1
dj-database-url==0.5.0
Django==4.1.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.5.0
django-storages==1.13.1
django-versatileimagefield==2.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.4
paypalrestsdk==1.13.1
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
python-magic==0.4.27
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
whitenoise==6.2.0
